🔧 Fix compile errors batch 1: SELinux, MagiskPatcher, DeviceBehavior, etc.
Fixed compilation errors found via local GCC build:

1. OEMDeepSpoofing.h: Added oemId/oemBrand to MIUIConfig struct
2. OEMDeepSpoofing.cpp: const QMap → .value() for const-safe access
3. EmulatorDetectionBypass.cpp: Removed duplicate destructor
4. EmulatorDetectionBypass.cpp: Added PlayIntegrityManager include
5. HypervisorBypass.cpp: Fixed literal.toStdString() pattern
6. MagiskPatcher.h/cpp: Added missing s_instance, methods
7. SELinuxManager.cpp: SELinuxConfig → MitigationSELinuxConfig everywhere
8. SELinuxManager.cpp: Fixed typo contextes → contexts
9. UniqueDeviceGenerator.cpp: Fixed ambiguous bounded() calls
10. DeviceBehaviorManager.h: Added missing s_instance
11. SafetyNetAdvancedBypass.cpp: Fixed 2 parenthesis errors
12. CryptoUtils.cpp: Fixed wrong include path
13. openssl_stub.h: Added RAND_bytes/RAND_seed stubs
14. PlayIntegrityManager: Added QJsonArray, brand field, removed duplicate dtor
15. 7+ files: Fixed wrong #include paths

+++ src/ReDroidController/UniqueDeviceGenerator.cpp
@@ -151,7 +151,7 @@ quint32 UniqueDeviceGenerator::getSecureRandomUInt32(quint32 min, quint32 max) c
                         (static_cast<quint32>(bytes[3]) << 24);
         } else {
             // Fallback to Qt's random generator
-            randomValue = QRandomGenerator::global()->bounded(0, UINT32_MAX);
+            randomValue = QRandomGenerator::global()->bounded(static_cast<quint32>(0), static_cast<quint32>(UINT32_MAX));
         }
     } while (randomValue >= limit);
     
@@ -172,7 +172,7 @@ quint64 UniqueDeviceGenerator::getSecureRandomUInt64() const {
     }
     
     // Fallback
-    return QRandomGenerator::global()->bounded(0, INT64_MAX);
+    return QRandomGenerator::global()->bounded(static_cast<quint64>(0), static_cast<quint64>(INT64_MAX));
 }
 
 QString UniqueDeviceGenerator::getSecureRandomHex(size_t length) const {
